extract page templates (generateDocumentJSON, etc)
towards basic dynamcis pt. 1 (#124)
